Replace presaleStart & End with refundStart.
presaleStart and presaleEnd are not really very useful. I believe this
contract should not have a 'start block' instead it should allow
investment from the very beginning. I don't see a good reason to have a
presale end. It might create a 'rush' for people invest early, but I
believe it is more useful to have anyone allow to invest at anytime.

The refundStartBlock can be used to 'lock' people into the contract
until a certain point. This may be a useful mechanic to stop whales from
filling up the cap, then refunding immediately before the token sale.